Initial implementation for a million of devices
